feat(trouble): add de-duplication for LSP locations
When multiple LSP clients are connected to the same buffer, a request
for locations (e.g. "lsp references") will ask *both* clients. The
tsserver and angularls are exceptionally annoying in that they overlap
in functionality, but not enough to use only one of them for Angular TS
component files.

This is a workaround where we simply duplicate locations before
displaying the results. This is done via a 'monkey patch' approach where
we just rewrite part of the trouble plugin (see
plugins.config.trouble).