fix: exclude go and rust from mason autoinstall
These LSPs seem to change their versioning based on the project they're being opened in. For example, the same gopls binary reports a different compilation Go version depending on the version of Go I'm using in the project.
To address this, we just turn off auto install for these LSPs so we just let the users provide their own LSP.